+++ Changelog.md
@@ -1,5 +1,11 @@
 # Vanilo Workflow Changelog
 
+## Unreleased
+##### 2023-XX-YY
+
+- Added support for callback hook on transition for saving the subject
+- Added support for passing parameters to the transition
+
 ## 1.0.0
 ##### 2023-03-14